Connected groups fetch and form to navbar
